@@ -10,39 +10,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.8",
-    "@angular/common": "^18.2.8",
-    "@angular/compiler": "^18.2.8",
-    "@angular/core": "^18.2.8",
-    "@angular/forms": "^18.2.8",
-    "@angular/platform-browser": "^18.2.8",
-    "@angular/platform-browser-dynamic": "^18.2.8",
-    "@angular/router": "^18.2.8",
-    "@asushares/core": "^2.5.0",
+    "@angular/animations": "^19.0.5",
+    "@angular/common": "^19.0.5",
+    "@angular/compiler": "^19.0.5",
+    "@angular/core": "^19.0.5",
+    "@angular/forms": "^19.0.5",
+    "@angular/platform-browser": "^19.0.5",
+    "@angular/platform-browser-dynamic": "^19.0.5",
+    "@angular/router": "^19.0.5",
+    "@asushares/core": "^2.7.2",
     "@popperjs/core": "^2.11.8",
     "@types/uuid": "^10.0.0",
     "bootstrap": "^5.3.3",
     "bootstrap-icons": "^1.11.3",
     "jquery": "^3.7.1",
     "rxjs": "~7.8.1",
-    "tslib": "^2.8.0",
-    "uuid": "^10.0.0",
-    "zone.js": "<0.15.0"
+    "tslib": "^2.8.1",
+    "uuid": "~10",
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.9",
-    "@angular/cli": "^18.2.9",
-    "@angular/compiler-cli": "^18.2.8",
+    "@angular-devkit/build-angular": "^19.0.6",
+    "@angular/cli": "^19.0.6",
+    "@angular/compiler-cli": "^19.0.5",
     "@types/bootstrap": "^5.2.10",
     "@types/fhir": "^0.0.41",
-    "@types/jasmine": "~5.1.4",
-    "@types/node": "^22.7.8",
-    "jasmine-core": "~5.4.0",
+    "@types/jasmine": "~5.1.5",
+    "@types/node": "^22.10.5",
+    "jasmine-core": "~5.5.0",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "<5.6"
+    "typescript": "~5.6"
   }
 }
