@@ -1,8 +1,14 @@
-import { bootstrapApplication } from '@angular/platform-browser';
-import { appConfig } from './app/app.config';
+import { bootstrapApplication, BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { importProvidersFrom } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { AppRoutingModule } from './app/app-routes.module';
 
-// importProvidersFrom(HttpClientModule);
-
-bootstrapApplication(AppComponent, appConfig)
+bootstrapApplication(AppComponent, {
+  providers: [
+    importProvidersFrom(AppRoutingModule, BrowserModule, FormsModule),
+    provideHttpClient(withInterceptorsFromDi())
+  ]
+})
   .catch((err) => console.error(err));
